Crate mnemonic

This is a Rust port of the mnemonic encoder originally written in C by Oren Tirosh and available from:

https://github.com/singpolyma/mnemonicode

These routines implement a method for encoding binary data into a sequence of words which can be spoken over the phone, for example, and converted back to data on the other side.

For more information, see:

http://web.archive.org/web/20101031205747/http://www.tothink.com/mnemonic/

Example

let bytes = [101, 2, 240, 6, 108, 11, 20, 97];

let s = mnemonic::to_string(&bytes);
assert_eq!(s, "digital-apollo-aroma--rival-artist-rebel");

let mut decoded = Vec::<u8>::new();
mnemonic::decode(s, &mut decoded).unwrap();

assert_eq!(decoded, [101, 2, 240, 6, 108, 11, 20, 97]);

Enums

Error

Errors returned by mnemonic decoding.

Constants

MN_FDEFAULT

Default format for encoding

Statics

MN_WORDS

Functions

decode

Decode the mnemonic string src into bytes, and write the bytes to dest.

encode

Encode the bytes of src into a mnemonic string, and write the string to dest

encode_with_format

Encode the bytes of s with a custom template.

to_string

Encode the bytes of src and return the results as a String

Type Aliases

Result

Result type returned by mnemonic decoding.